Remove bloc and let all be cubit
