+++ lib/ui/signin/signin_screen.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_template/ui/extensions/context_extensions.dart';
 import 'package:flutter_template/ui/section/error_handler/error_handler_cubit.dart';
-import 'package:flutter_template/ui/signin/signin_bloc.dart';
+
+import 'package:flutter_template/ui/signin/signin_cubit.dart';
 
 class SignInScreen extends StatelessWidget {
   const SignInScreen({Key? key}) : super(key: key);
@@ -10,7 +11,7 @@ class SignInScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocProvider(
-      create: (context) => SignInBloc(context.read<ErrorHandlerCubit>()),
+      create: (context) => SignInCubit(context.read<ErrorHandlerCubit>()),
       child: _SignInContentScreen(),
     );
   }
@@ -19,7 +20,7 @@ class SignInScreen extends StatelessWidget {
 class _SignInContentScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<SignInBloc, SignInState>(
+    return BlocBuilder<SignInCubit, SignInBaseState>(
       builder: (context, state) {
         return Scaffold(
           appBar: AppBar(
@@ -29,12 +30,11 @@ class _SignInContentScreen extends StatelessWidget {
             mainAxisSize: MainAxisSize.min,
             children: [
               Expanded(child: _SignInForm()),
-              if (context.read<SignInBloc>().state.error.isNotEmpty)
+              if (context.read<SignInCubit>().state.error.isNotEmpty)
                 Text(context.localizations
-                    .error(context.read<SignInBloc>().state.error)),
+                    .error(context.read<SignInCubit>().state.error)),
               TextButton(
-                  onPressed: () =>
-                      context.read<SignInBloc>().add(SignInEvent.login()),
+                  onPressed: () => context.read<SignInCubit>().login(),
                   child: Text(context.localizations.sign_in))
             ],
           ),
@@ -52,7 +52,7 @@ class _SignInForm extends StatefulWidget {
 class _SignInFormState extends State<_SignInForm> {
   final _emailTextController = TextEditingController();
   final _passwordTextController = TextEditingController();
-  late SignInBloc _signInBlock;
+  late SignInCubit _signInCubit;
 
   @override
   void dispose() {
@@ -64,10 +64,10 @@ class _SignInFormState extends State<_SignInForm> {
   @override
   void initState() {
     super.initState();
-    _signInBlock = context.read<SignInBloc>();
+    _signInCubit = context.read<SignInCubit>();
     // TODO: This should be bound
-    _emailTextController.text = _signInBlock.state.email ?? '';
-    _passwordTextController.text = _signInBlock.state.password ?? '';
+    _emailTextController.text = _signInCubit.state.email ?? '';
+    _passwordTextController.text = _signInCubit.state.password ?? '';
   }
 
   @override
@@ -80,8 +80,7 @@ class _SignInFormState extends State<_SignInForm> {
           child: Container(
             child: TextField(
               controller: _emailTextController,
-              onChanged: (String text) =>
-                  _signInBlock.add(SignInEvent.changeEmail(email: text)),
+              onChanged: (String text) => _signInCubit.changeEmail(text),
               decoration: InputDecoration(
                 border: OutlineInputBorder(),
                 labelText: context.localizations.mail,
@@ -95,8 +94,8 @@ class _SignInFormState extends State<_SignInForm> {
               child: TextField(
             obscureText: true,
             controller: _passwordTextController,
-            onChanged: (String password) => _signInBlock
-                .add(SignInEvent.changePassword(password: password)),
+            onChanged: (String password) =>
+                _signInCubit.changePassword(password),
             decoration: InputDecoration(
               border: OutlineInputBorder(),
               labelText: context.localizations.password,
